[refactor] refactor user module's controller and services
Now you can add apis easier and more convenient.
