+++ controller/user.go
@@ -2,12 +2,12 @@ package controller
 
 import (
 	"OpenTeens/model"
+	"OpenTeens/services"
 	"OpenTeens/utils"
 	"fmt"
 	"github.com/gin-gonic/gin"
 	"strconv"
 	"strings"
-	"time"
 )
 
 // 代码命名格式：
@@ -18,7 +18,7 @@ import (
 
 // IndexHandler 处理访问Root的请求
 func IndexHandler(c *gin.Context) {
-	c.JSON(200, gin.H{"code": 200, "msg": "Hello, Tsinglan!", "data": gin.H{"token": utils.GenerateToken(), "verify": utils.GenerateRandomNumber()}})
+	c.JSON(200, gin.H{"code": 200, "msg": "Hello, OpenTeens!", "data": gin.H{"token": utils.GenerateToken(), "verify": utils.GenerateRandomNumber()}})
 }
 
 // UserSendEmailHandler 发送邮箱验证码的处理
@@ -30,33 +30,13 @@ func UserSendEmailHandler(c *gin.Context) {
 	// 		-> 发送邮件 -> FuncUserEmailSend（会调用数据库）
 	// -> 返回错误信息
 
-	err := UserSendEmailService(c.PostForm("email"), c.ClientIP())
+	err := services.UserSendEmailService(c.PostForm("email"), c.ClientIP())
 	// 如果返回值是Error，说明发送失败，返回400，否则返回200
 	if err != true {
 		c.JSON(400, gin.H{"msg": err})
 		return
 	}
 	c.JSON(200, gin.H{"msg": "Code send successfully!"})
-}
-
-// UserSendEmailService 发送邮箱验证码业务逻辑（包括存入数据库）
-func UserSendEmailService(email string, ipaddr string) any {
-	emailCheck := utils.ToolValidateValue(email, utils.Email)
-	serviceCheck := FuncUserEmailServiceCheck(ipaddr, email) // TODO: Need to be changed
-
-	if emailCheck && serviceCheck {
-		code := utils.GenerateRandomNumber()
-		FuncUserEmailSend(email, code, ipaddr)
-		return true
-	} else if !emailCheck {
-		fmt.Println("==================> Send Email Failed to", email, "where ip is", ipaddr, "because of email is not valid.")
-		return "Email Pattern is not Valid."
-	} else if !serviceCheck {
-		fmt.Println("==================> Send Email Failed to", email, "where ip is", ipaddr, "because of service is not valid.")
-		return "Service is not valid. Usually you send too much email."
-	} else {
-		return "Unknown Error."
-	}
 }
 
 // UserVerifyEmailHandler 验证邮箱验证码的Handler
@@ -69,22 +49,13 @@ func UserVerifyEmailHandler(c *gin.Context) {
 		return
 	}
 
-	respond := UserVerifyEmailService(code, email)
+	respond := services.UserVerifyEmailService(code, email)
 	if respond == true {
 		c.JSON(200, gin.H{"msg": "Email Verify Successfully!"})
 	} else {
 		c.JSON(400, gin.H{"msg": respond})
 	}
 
-}
-
-func UserVerifyEmailService(code string, email string) any {
-	// 从数据库中获取code和email
-	success := model.DBGetEmailVerifyByCodeAndEmail(code, email)
-	if !success {
-		return "Code is not valid."
-	}
-	return true
 }
 
 // UserEmailExistHandler 检查邮箱是否已经注册
@@ -175,40 +146,12 @@ func UserLoginHandler(c *gin.Context) {
 		c.JSON(400, gin.H{"code": 400, "msg": "Username is not valid."})
 		return
 	}
-	success, token := UserLoginService(username, password)
+	success, token := services.UserLoginService(username, password)
 	if success == true {
 		c.JSON(200, gin.H{"code": 200, "msg": "Login Successfully!", "data": token})
 	} else {
 		c.JSON(400, gin.H{"code": 400, "msg": "Login Failed.", "data": false})
 	}
-}
-
-// UserLoginService 用户登录业务逻辑（包括存入数据库）
-func UserLoginService(username string, password string) (bool, string) {
-	//	先加密密码
-	passwordHashed, _ := utils.HashPassword(password)
-
-	// 然后首先用用户名密码DBUserCheckAccountByUsernameAndPasswordHashed
-	ok, user := model.DBUserCheckAccount(username, passwordHashed, model.LoginByUsername)
-	if ok {
-		_, token := model.DBUserLoginCreateToken(user, "username")
-		return true, token
-	}
-
-	// 然后用邮箱密码DBUserCheckAccountByEmailAndPasswordHashed
-	ok, user = model.DBUserCheckAccount(username, passwordHashed, model.LoginByEmail)
-	if ok {
-		_, token := model.DBUserLoginCreateToken(user, "email")
-		return true, token
-	}
-
-	//	然后用uid密码DBUserCheckAccountByUidAndPasswordHashed
-	ok, user = model.DBUserCheckAccount(username, passwordHashed, model.LoginByUserID)
-	if ok {
-		_, token := model.DBUserLoginCreateToken(user, "uid")
-		return true, token
-	}
-	return false, ""
 }
 
 // UserGetInfoHandler 获取用户信息
@@ -278,27 +221,12 @@ func UserSiteMessageSendHandler(c *gin.Context) {
 		messageData.Recipient = append(messageData.Recipient, uint(recipientID))
 	}
 
-	success := UserSiteMessageSendService(messageData.SenderID, messageData.Content, messageData.Recipient)
+	success := services.UserSiteMessageSendService(messageData.SenderID, messageData.Content, messageData.Recipient)
 	if success {
 		c.JSON(200, gin.H{"msg": "Message sent successfully!", "data": true})
 	} else {
 		c.JSON(400, gin.H{"msg": "Failed to send message", "data": false})
 	}
-}
-
-// UserSiteMessageSendService 发送站内信业务逻辑
-func UserSiteMessageSendService(senderID uint, content string, recipients []uint) bool {
-	// 创建站内信
-	message := model.SiteMessage{SenderID: senderID, Content: content, SentAt: time.Now()}
-	model.DBUserSiteMessageCreate(&message)
-
-	// 为每个接收者创建记录
-	for _, receiverID := range recipients {
-		recipient := model.SiteMessageRecipient{MessageID: message.ID, ReceiverID: receiverID}
-		model.DBUserSiteMessageRecipientCreate(&recipient)
-	}
-
-	return true // 或者根据实际情况返回成功与否
 }
 
 // UserSiteMessageGetAllHandler 获取当前用户所有站内信
